module_5/requirements.txt
beautifulsoup4
flask
huggingface_hub
llama-cpp-python
psycopg[binary]
pytest
pytest-cov
pytest-mock
sphinx
sphinx-rtd-theme
urllib3
